Update makedots.py
Added a second conn.commit() statement.  Without it, the code will not commit to hard disk the last features of the shapefiles.  Under the old code, results are written to disk only every thousand iterations, meaning that up to 999 features in a shapefile could be lost.  

This edit also ensures that the people.db file has the correct population totals.  For instance, a U.S. only run will result in 308,745,538 observations, the total U.S. resident population in the 2010 Census.